adding utils, routers, and controller to BE
