+++ backend/middleware.py
@@ -1,9 +1,14 @@
+from fastapi.middleware.cors import CORSMiddleware
 from starlette import status
 from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
 from starlette.requests import Request
 from starlette.responses import Response
 from starlette.types import ASGIApp
 
+origins = [
+    "http://localhost:8000",
+    "http://localhost:3000",
+]
 
 class LimitUploadSize(BaseHTTPMiddleware):
     def __init__(self, app: ASGIApp, max_upload_size: int) -> None:
@@ -18,3 +23,15 @@ class LimitUploadSize(BaseHTTPMiddleware):
             if content_length > self.max_upload_size:
                 return Response(status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE)
         return await call_next(request)
+    
+# REGISTER MIDDLEWARE
+def initMiddleware(app: ASGIApp) -> None:
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=origins,
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+
+    app.add_middleware(LimitUploadSize, max_upload_size=10_000_000) #10mb
